Allow motor test to use channel order
MAV_CMD_DO_MOTOR_TEST expects the index in the first parameter to be a
"test order" that varies by vehicle but maintains a defined (clockwise)
order.  However for people assembling the drone and for ground control
stations it may be easier to just use the channel order, since the
numbers are very easy to get and it's possible to maintain consistency
across vehicles.